Restructure tabs.
