+++ index-v6.js
@@ -1,2 +1,3 @@
 export { Modal } from './components/modal';
 export { Button, SegmentedButtonGroup } from './components/button';
+export { Tab } from './components/tabs';
